requirements.txt
discord.py
Pillow
psutil
git+https://github.com/oliver-ni/Flag-Parsing.git
pytz
frozendict
numpy
orjson
python-dotenv